requirements.txt
flask
requests
pyopenssl
ndg-httpsclient
pyasn1
pyaml
bcrypt
certifi
cffi
charset-normalizer
colorama
cryptography
future
idna
ipaddress
mypy-extensions
netmiko
nornir
nornir-jinja2
nornir-netbox
nornir-netmiko
nornir-utils
nornir-scrapli
ntc-templates
paramiko
pycparser
PyNaCl
pyserial
python-netbox
requests
ruamel.yaml
ruamel.yaml.clib
scp
six
tenacity
textfsm
ttp
typing-extensions
urllib3
genie[full]
pyats[full]
pyadaptivecards
pynetbox
python-slugify
